refactor: split AlertQueueRepository out of AlertQueue
Data access moves into AlertQueueRepository, following the codebase's
repository convention; AlertQueue keeps the alert-identity/batching
logic and delegates I/O. The repository is provided but not exported
from DispatchModule, so AlertQueue stays the module's only public
seam — addresses PR review feedback on #61.

Updates the design spec's placement diagram, interface, and Phase 3
test wiring to match.
