+++ apps/scrubjay-discord/src/features/dispatch/dispatch.module.ts
@@ -1,11 +1,12 @@
 import { Module } from "@nestjs/common";
 import { DiscordModule } from "@/discord/discord.module";
 import { AlertQueue } from "./alert-queue";
+import { AlertQueueRepository } from "./alert-queue.repository";
 import { EBirdDispatcherService } from "./ebird-dispatcher.service";
 
 @Module({
   exports: [AlertQueue, EBirdDispatcherService],
   imports: [DiscordModule],
-  providers: [AlertQueue, EBirdDispatcherService],
+  providers: [AlertQueue, AlertQueueRepository, EBirdDispatcherService],
 })
 export class DispatchModule {}
